Move getTracksFromPlaylist
